Change grandparent current post view to route (not modal)
