+++ src/Routes.tsx
@@ -20,6 +20,7 @@ import FamilyMsgView from "./shared/features/FamilyMsgView/FamilyMsgView";
 import CreatePost from "./shared/features/CreatePost/CreatePost";
 import { Grid } from "@material-ui/core";
 import NewGrandparentReply from "./shared/features/GrandparentViews/GrandparentReply/NewGrandparentReply";
+import GrandparentCurrentPost from "./shared/features/GrandparentViews/Inbox/GrandparentCurrentPost";
 
 interface IRoutes {
   isLoggedIn: boolean;
@@ -70,6 +71,7 @@ const Routes: React.FC<IRoutes & RouteComponentProps> = (props) => {   // {} is 
               <ProtectedRouteHoc exact setIsLoading={setIsLoading} path="/posts" isLoggedIn={isLoggedIn} public={false} RouteComponent={PostsView} />
               <ProtectedRouteHoc exact path="/postDetails" isLoggedIn={isLoggedIn} public={false} RouteComponent={FamilyMsgView} />
               <ProtectedRouteHoc exact path="/newPost" isLoggedIn={isLoggedIn} public={false} RouteComponent={CreatePost} />
+              <ProtectedRouteHoc exact path="/currentPost" isLoggedIn={isLoggedIn} public={false} RouteComponent={GrandparentCurrentPost} />
               <ProtectedRouteHoc exact path="/newReply" isLoggedIn={isLoggedIn} public={false} RouteComponent={NewGrandparentReply} />
               <ProtectedRouteHoc exact setIsLoading={setIsLoading} path="/settings" isLoggedIn={isLoggedIn} public={false} RouteComponent={SettingsView} />
               <ProtectedRouteHoc exact path="/photoReplyPrototype" isLoggedIn={isLoggedIn} public={false} RouteComponent={PhotoReplyPrototype} />
